Production

+++ src/controllers/users.controller.js
@@ -1,76 +1,177 @@
 import db from "../models/index.js";
 const Users = db.users;
+const Referrals = db.referrals;
 import bcrypt from "bcrypt";
+import jwt from "jsonwebtoken";
+import adminCheck from "../services/admincheck.service.js";
 
-// Done
+import mongoose from "mongoose";
+const ObjectId = mongoose.Types.ObjectId;
+
+// Fetch all users - Admin Only (need to be updated)
 const findAll = (req, res) => {
-  Users.find()
+  const { admin, basic, pro } = req.query;
+
+  var condition = {};
+
+  if (admin) {
+    condition = { "type.isAdmin": true };
+  } else if (basic) {
+    condition = { "type.accountType.member": "Basic Member" };
+  } else if (pro) {
+    condition = { "type.accountType.member": "Pro Member" };
+  } else {
+    condition = {};
+  }
+
+  Users.find(condition)
     .sort({ createdAt: -1 })
+    .populate({
+      path: "referral",
+      select:
+        "referralCode referralCount referralAccount referralTotalAmount referralAvailableAmount referralWithDraw referralWithDrawBank referralWithDrawHistory referralStatus ",
+    })
     .then((result) => {
-      res.send({
+      if (!result) {
+        return res.status(404).send({
+          message: "No users found!",
+        });
+      }
+
+      const data = result.map((item) => {
+        var {
+          _id,
+          name,
+          username,
+          email,
+          phone,
+          address,
+          birthday,
+          type,
+          referral,
+        } = item;
+
+        birthday = new Date(birthday).toString();
+
+        return {
+          id: _id,
+          name,
+          username,
+          email,
+          phone,
+          address,
+          birthday,
+          memberType: type.accountType.member,
+          subscription: {
+            startAt: new Date(type.accountType.subscription.startAt).toString(),
+            expiredAt: new Date(
+              type.accountType.subscription.expiredAt
+            ).toString(),
+          },
+          isNew: type.accountType.isNew,
+          adminType: type.isAdmin,
+          isActivated: type.isActivated,
+          referral,
+        };
+      });
+
+      res.status(200).send({
         message: "Users fetched successfully!",
-        timestamp: new Date().toString(),
-        data: result,
+        data,
       });
     })
     .catch((err) => {
+      console.log(err);
       return res.status(500).send({
         message: err.message || "Some error occurred while fetching the Users.",
       });
     });
 };
 
-// Done
+// Find a single user with an id (need to be updated)
 const findOne = (req, res) => {
   const { id } = req.params;
 
   if (!id) {
     return res.status(400).send({
-      message: "User ID is required.",
+      message: "User ID is required!",
     });
   }
 
   Users.findById(id)
+    .populate({
+      path: "referral",
+      select:
+        "referralCode referralCount referralAccount referralTotalAmount referralAvailableAmount referralWithDraw referralWithDrawBank referralWithDrawHistory referralStatus ",
+    })
     .then((result) => {
       if (!result) {
         return res.status(404).send({
-          message: "User not found",
+          message: "User not found!",
         });
       }
 
-      res.send({
-        message: "User fetched successfully!",
-        timestamp: new Date().toString(),
-        data: {
-          name: result.name,
-          username: result.username,
-          email: result.email,
-          image: {
-            imageName: result.image.imageName,
-            imageLink: result.image.imageLink,
-          },
-          type: {
-            memberType: result.type.accountType.member,
-            isNew: result.type.accountType.isNew,
-            adminType: result.type.isAdmin,
-          },
-          referal: {
-            referalCode: result.referal.referalCode,
-            referalCount: result.referal.referalCount,
-            referalAccount: result.referal.referalAccount,
+      const {
+        _id,
+        name,
+        username,
+        email,
+        phone,
+        address,
+        birthday,
+        image,
+        type,
+        referral,
+      } = result;
+
+      if (type.accountType.subscription.expiredAt == 0) {
+        type.accountType.subscription.expiredAt = null;
+      } else {
+        type.accountType.subscription.expiredAt = new Date(
+          type.accountType.subscription.expiredAt
+        ).toString();
+      }
+
+      const data = {
+        id: _id,
+        name,
+        username,
+        email,
+        phone,
+        address,
+        birthday: new Date(birthday).toString(),
+        image,
+        type: {
+          isAdmin: type.isAdmin,
+          isActivated: type.isActivated,
+          accountType: {
+            member: type.accountType.member,
+            subscription: {
+              startat: new Date(
+                type.accountType.subscription.startAt
+              ).toString(),
+              expiredAt: type.accountType.subscription.expiredAt,
+            },
+            isNew: type.accountType.isNew,
           },
         },
+        referral,
+      };
+
+      res.status(200).send({
+        message: "User fetched successfully!",
+        data,
       });
     })
     .catch((err) => {
       return res.status(500).send({
-        message: err.message || "Some error while retrieving the user.",
+        message: err.message || "Some error while retrieving the user!",
       });
     });
 };
 
-// Done
-const deleteUSer = (req, res) => {
+// Delete a user with the specified id in the request - Admin Only (Done)
+const deleteUSer = async (req, res) => {
   const { id } = req.params;
 
   if (!id) {
@@ -79,16 +180,25 @@ const deleteUSer = (req, res) => {
     });
   }
 
-  Users.findByIdAndRemove(id)
+  // If selected user is admin, then don't delete
+  const admincheck = await adminCheck(id);
+
+  if (admincheck) {
+    return res.status(405).send({
+      message: "Admin user can't be deleted!",
+    });
+  }
+
+  await Users.findByIdAndRemove(id)
     .then((result) => {
       if (!result) {
         return res.status(404).send({
           message: "User not found",
         });
       }
+
       res.send({
-        message: "User deleted successfully.",
-        timestamp: new Date().toString(),
+        message: "User deleted successfully!",
       });
     })
     .catch((err) => {
@@ -98,7 +208,7 @@ const deleteUSer = (req, res) => {
     });
 };
 
-// Done
+// Update a user by the id in the request (Done)
 const update = (req, res) => {
   const { id } = req.params;
 
@@ -115,28 +225,27 @@ const update = (req, res) => {
           message: "User not found",
         });
       }
-      res.send({
-        message: "User updated successfully.",
-        timestamp: new Date().toString(),
-        data: {
+
+      const token = jwt.sign(
+        {
+          id: result.id,
+          email: result.email,
           name: result.name,
           username: result.username,
-          email: result.email,
-          image: {
-            imageName: result.image.imageName,
-            imageLink: result.image.imageLink,
-          },
-          type: {
-            memberType: result.type.accountType.member,
-            isNew: result.type.accountType.isNew,
-            adminType: result.type.isAdmin,
-          },
-          referal: {
-            referalCode: result.referal.referalCode,
-            referalCount: result.referal.referalCount,
-            referalAccount: result.referal.referalAccount,
-          },
+          admin: result.type.isAdmin,
+          role: result.type.accountType.member,
+          isActivated: result.type.isActivated,
+          image: result.image.imageLink,
         },
+        process.env.JWT_SECRET,
+        {
+          expiresIn: "12h",
+        }
+      );
+
+      res.send({
+        message: "User updated successfully.",
+        token,
       });
     })
     .catch((err) => {
@@ -146,7 +255,7 @@ const update = (req, res) => {
     });
 };
 
-// Done
+// Change password (Done)
 const changePassword = (req, res) => {
   const { id } = req.params;
   const { oldPassword } = req.body;
@@ -160,7 +269,7 @@ const changePassword = (req, res) => {
 
   if (newPassword === oldPassword) {
     return res.status(409).send({
-      message: "New Password cannot be the same as Old Password",
+      message: "New Password cannot be the same as Old Password!",
     });
   }
 
@@ -200,9 +309,9 @@ const changePassword = (req, res) => {
                     message: "User not found",
                   });
                 }
+
                 res.send({
                   message: "User's password updated successfully.",
-                  timestamp: new Date().toString(),
                 });
               })
               .catch((err) => {
@@ -224,7 +333,7 @@ const changePassword = (req, res) => {
     });
 };
 
-// Done
+// Change user's image (Done)
 const changeProfilePicture = (req, res) => {
   const { id } = req.params;
 
@@ -268,9 +377,27 @@ const changeProfilePicture = (req, res) => {
           message: "User not found",
         });
       }
+
+      const token = jwt.sign(
+        {
+          id: result.id,
+          email: result.email,
+          name: result.name,
+          username: result.username,
+          admin: result.type.isAdmin,
+          role: result.type.accountType.member,
+          isActivated: result.type.isActivated,
+          image: result.image.imageLink,
+        },
+        process.env.JWT_SECRET,
+        {
+          expiresIn: "12h",
+        }
+      );
+
       res.send({
         message: "User's profile picture updated successfully.",
-        timestamp: new Date().toString(),
+        token,
       });
     })
     .catch((err) => {
@@ -282,6 +409,185 @@ const changeProfilePicture = (req, res) => {
     });
 };
 
+// Create Referral Code (Done)
+const createReferralCode = (req, res) => {
+  const { id } = req.params;
+  var { referralCode } = req.body;
+
+  if (!id) {
+    return res.status(400).send({
+      message: "User ID is required",
+    });
+  }
+
+  if (!referralCode)
+    // Generate random referral code with 8 characters
+    referralCode = Math.random().toString(36).substring(2, 10);
+
+  if (referralCode.length > 8) {
+    return res.status(400).send({
+      message: "Referral Code must be 8 characters or less",
+    });
+  }
+
+  Referrals.findOne({ referralCode: referralCode })
+    .then((result) => {
+      if (result) {
+        return res.status(409).send({
+          message: "Referral Code already exists",
+        });
+      }
+
+      const referral = new Referrals({
+        referralCode: referralCode,
+        referralUser: id,
+      });
+
+      referral
+        .save()
+        .then((result) => {
+          const referralId = result._id;
+
+          Users.findByIdAndUpdate(
+            id,
+            {
+              referral: referralId,
+            },
+            { new: true }
+          )
+            .then((result) => {
+              if (!result) {
+                return res.status(404).send({
+                  message: "User not found",
+                });
+              }
+
+              res.send({
+                message: "User's refer code created successfully.",
+              });
+            })
+            .catch((err) => {
+              return res.status(500).send({
+                message:
+                  err.message ||
+                  "Some error occurred while creating the referral code.",
+              });
+            });
+        })
+        .catch((err) => {
+          return res.status(500).send({
+            message:
+              err.message || "Some error occurred while creating the referral.",
+          });
+        });
+    })
+    .catch((err) => {
+      return res.status(500).send({
+        message:
+          err.message || "Some error occurred while creating the referral.",
+      });
+    });
+};
+
+// Change Pro Member into Basic Member While Pro Member's Subscription is Expired (Done - Not Tested)
+const changeProMemberToBasicMember = async (req, res) => {
+  const date = new Date().getTime();
+
+  await Users.updateMany(
+    { "type.accountType.subscription.expiredAt": { $lt: date } }, // Find all users whose subscription is expired
+    {
+      $set: {
+        "type.accountType.type": "Basic Member", // Change their account type to Basic Member
+        "type.accountType.subscription": {
+          // Set their subscription to null
+          expiredAt: null,
+        },
+      },
+    }
+  )
+    .then((result) => {
+      if (!result) {
+        return res.status(200).send({
+          message: "No user updated",
+        });
+      }
+
+      const count = result.nModified;
+
+      res.send({
+        message: `${count} user(s) changed from Pro Member to Basic Member.`,
+      });
+    })
+    .catch((err) => {
+      return res.status(500).send({
+        message:
+          err.message ||
+          "Some error occurred while changing Pro Member to Basic Member.",
+      });
+    });
+};
+
+// Request User Activation (Done)
+const requestUserActivation = async (req, res) => {
+  const { id } = req.params;
+
+  if (!id) {
+    return res.status(400).send({
+      message: "User ID is required",
+    });
+  }
+
+  const result = await Users.findById(id)
+    .then((result) => {
+      if (!result) {
+        return res.status(404).send({
+          message: "User not found",
+        });
+      }
+
+      return result;
+    })
+    .catch((err) => {
+      return res.status(500).send({
+        message:
+          err.message || "Some error occurred while retrieving the user.",
+      });
+    });
+
+  if (result.type.isActivated === true) {
+    return res.status(409).send({
+      message: "User already activated",
+    });
+  }
+
+  // Generate token
+  const token = jwt.sign(
+    {
+      id: result._id,
+      name: result.name,
+      username: result.username,
+      email: result.email,
+    },
+    process.env.JWT_SECRET,
+    {
+      expiresIn: "10m",
+    }
+  );
+
+  // Send email
+  const response = await signupMailer(result.email, token);
+
+  if (response == "Email sent") {
+    return res.status(200).send({
+      message: "User registered successfully! Please check your email.",
+    });
+  } else {
+    return res.status(500).send({
+      message: "Failed to send email.",
+    });
+  }
+};
+
 export {
   findAll,
   findOne,
@@ -289,4 +595,7 @@ export {
   update,
   changePassword,
   changeProfilePicture,
+  createReferralCode,
+  changeProMemberToBasicMember,
+  requestUserActivation,
 };
